refactor(console): rearrange creation components to get "get-started, V2, V4" in dedicated folder
