+++ gravitee-apim-console-webui/src/management/api/_api.scss
@@ -1,6 +1,6 @@
 @import 'analytics/logs/logs';
 @import 'analytics/overview/analytics';
-@import 'creation/steps/api-creation';
+@import 'creation-v2/steps/api-creation';
 @import 'audit/history/apiHistory';
 @import 'design/policies/policy';
 @import 'proxy/properties-v1/properties';
