+++ gravitee-apim-console-webui/src/management/api/creation-get-started/api-creation-get-started.component.html
