+++ gravitee-apim-console-webui/src/management/api/creation-v2/steps/_api-creation.scss
