+++ gravitee-apim-console-webui/src/management/api/creation-v2/steps/api-creation-step3.html
