+++ gravitee-apim-console-webui/src/management/api/creation-v2/steps/form.helper.ts
